Update orientation calculation for landscape modes
The code changes in `VisionCameraFaceDetectorPlugin.kt` update the orientation calculation for
landscape modes. The `Orientation.LANDSCAPE_LEFT` now returns 270 degrees, and `Orientation.LANDSCAPE_RIGHT`
returns 90 degrees. This ensures that the correct orientation is applied when capturing images in landscape
modes.

For the second argument rotationDegrees in MLKit's InputImage.fromMediaImage(), you need to provide
the angle at which the image is rotated, not the angle by which to rotate it.